@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
@@ -30,17 +30,17 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="BoltOn" Version="0.11.0" />
-    <PackageReference Include="BoltOn.Data.EF" Version="0.7.0" />
-    <PackageReference Include="BoltOn.Data.CosmosDb" Version="0.4.0" />
-    <PackageReference Include="BoltOn.Bus.MassTransit" Version="0.4.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.2.0" />
+    <PackageReference Include="BoltOn" Version="0.13.0" />
+    <PackageReference Include="BoltOn.Data.EF" Version="0.9.0" />
+    <PackageReference Include="BoltOn.Data.CosmosDb" Version="0.5.0" />
+    <PackageReference Include="BoltOn.Bus.MassTransit" Version="0.6.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="3.1.1" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.1" />
     <PackageReference Include="MassTransit.Extensions.DependencyInjection" Version="6.0.2" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="2.2.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="3.1.1" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="3.1.1" />
     <PackageReference Include="MassTransit.RabbitMQ" Version="6.0.2" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.2.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.1" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\BoltOn.Samples.Application\BoltOn.Samples.Application.csproj" />
