@@ -1,22 +1,17 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App">
-      <PrivateAssets Condition="'%(PackageReference.Version)' == ''">all</PrivateAssets>
-      <Publish Condition="'%(PackageReference.Version)' == ''">true</Publish>
-    </PackageReference>
-    <PackageReference Include="BoltOn" Version="0.11.0" />
-    <PackageReference Include="BoltOn.Data.EF" Version="0.7.0" />
-    <PackageReference Include="BoltOn.Data.CosmosDb" Version="0.4.0" />
-    <PackageReference Include="BoltOn.Bus.MassTransit" Version="0.4.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.2.0" />
+    <PackageReference Include="BoltOn" Version="0.13.0" />
+    <PackageReference Include="BoltOn.Data.EF" Version="0.9.0" />
+    <PackageReference Include="BoltOn.Data.CosmosDb" Version="0.5.0" />
+    <PackageReference Include="BoltOn.Bus.MassTransit" Version="0.6.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.1" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.1" />
     <PackageReference Include="MassTransit.Extensions.DependencyInjection" Version="6.0.2" />
     <PackageReference Include="MassTransit.RabbitMQ" Version="6.0.2" />
   </ItemGroup>
