@@ -10,6 +10,7 @@ using BoltOn.Samples.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
 using MassTransit;
 using System;
+using Microsoft.Extensions.Hosting;
 
 namespace BoltOn.Samples.WebApi
 {
@@ -24,7 +25,7 @@ namespace BoltOn.Samples.WebApi
 
 		public void ConfigureServices(IServiceCollection services)
 		{
-			services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+			services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
 			services.BoltOn(options =>
 			{
 				options.BoltOnEFModule();
@@ -62,10 +63,15 @@ namespace BoltOn.Samples.WebApi
 
         }
 
-		public void Configure(IApplicationBuilder app, IApplicationLifetime appLifetime)
+		public void Configure(IApplicationBuilder app, IHostApplicationLifetime appLifetime)
 		{
-			app.UseMvc();
-			app.ApplicationServices.TightenBolts();
+			app.UseRouting(); 
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapRazorPages();
+                endpoints.MapDefaultControllerRoute();
+            });
+            app.ApplicationServices.TightenBolts();
 			appLifetime.ApplicationStopping.Register(() => app.ApplicationServices.LoosenBolts());
 		}
 	}
